Include print styles in main stylesheet
No need to do this separately which makes an extra http request, plus
this has the benefit of delaying print styles until needed. I’ve been
doing this for a long time now and as its been well supported to
include print styles using a media block. The HTML5 Boilerplate has
stressed this as a good default approach for many years now.
https://github.com/h5bp/html5-boilerplate/blob/master/src/css/main.css#L
205-L209